requirements.txt
aniso8601==8.1.1
attrs==20.3.0
click==7.1.2
coverage==5.4
Flask==1.1.2
Flask-RESTful==0.3.8
iniconfig==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
packaging==20.9
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
SQLAlchemy==1.3.23
toml==0.10.2
Werkzeug==1.0.1
